[PUX-147] Refactor models.
