+++ src/clients/authentication-client.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import Cache from 'node-cache';
-import { AuthenticationResponse } from 'models/authentication';
+import { AuthenticationResponse } from 'models/payments-api/responses';
 import logger from 'middleware/logger';
 import config from 'config';
 import { HttpException } from 'middleware/errors';
