+++ src/models/payments-api/requests/authentication.ts
@@ -1,0 +1,8 @@
+/* eslint-disable camelcase */
+
+export interface AuthenticationRequest {
+  grant_type: string;
+  client_id: string;
+  client_secret: string;
+  scope: string;
+}
