+++ src/models/payments-api/responses/authentication.ts
@@ -5,10 +5,3 @@ export interface AuthenticationResponse {
   expires_in: number;
   token_type: string;
 }
-
-export interface AuthenticationRequest {
-  grant_type: string;
-  client_id: string;
-  client_secret: string;
-  scope: string;
-}
