+++ src/models/payments-api/responses/index.ts
@@ -1,0 +1,3 @@
+export * from './authentication';
+export * from './single_immediate_payment';
+export * from './single_immediate_provider';
