+++ src/models/payments/request.ts
@@ -2,29 +2,9 @@
 
 import config from 'config';
 import { v4 as uuid } from 'uuid';
-
-import Participant from './participant';
-import References from './references';
-import { SupportedCurrency } from './response';
-import { AuthFlowRedirectRequest } from './auth-flow';
-
-interface SingleImmediatePayment {
-  single_immediate_payment_id: string;
-  provider_id: string;
-  scheme_id?: string;
-  fee_option_id?: string;
-  currency: SupportedCurrency;
-  amount_in_minor: number;
-  beneficiary: Participant;
-  remitter?: Participant;
-  references: References;
-}
-
-interface PaymentApiRequest {
-  auth_flow: AuthFlowRedirectRequest;
-  single_immediate_payment: SingleImmediatePayment;
-  webhook_uri?: string | null;
-}
+import { SupportedCurrency } from 'models/payments-api/common';
+import { SingleImmediatePaymentRequest } from 'models/payments-api/requests';
+import { isTypeOf } from 'utils';
 
 export interface PaymentRequest {
   provider_id: string;
@@ -35,14 +15,28 @@ export interface PaymentRequest {
   payment_id?: string;
 }
 
-export const buildPaymentApiRequest = ({
+// Dynamic type check to validate request.
+export const isPaymentRequest = (obj: any): obj is PaymentRequest => {
+  const invalidParam = [
+    isTypeOf(obj.provider_id, ['string']),
+    isTypeOf(obj.currency, ['string', 'undefined']),
+    isTypeOf(obj.amount_in_minor, ['number', 'undefined']),
+    isTypeOf(obj.reference, ['string', 'undefined']),
+    isTypeOf(obj.scheme_id, ['string', 'undefined']),
+    isTypeOf(obj.payment_id, ['string', 'undefined'])
+  ].some(isType => !isType);
+
+  return !invalidParam;
+};
+
+export const intoSingleImmediatePaymentRequest = ({
   scheme_id,
   provider_id,
   currency = 'GBP',
   amount_in_minor = 1,
   reference = 'Test Payment',
   payment_id = uuid()
-}: PaymentRequest): PaymentApiRequest => ({
+}: PaymentRequest): SingleImmediatePaymentRequest => ({
   single_immediate_payment: {
     single_immediate_payment_id: payment_id,
     provider_id,
@@ -69,4 +63,4 @@ export const buildPaymentApiRequest = ({
   webhook_uri: config.WEBHOOK_URI || null
 });
 
-export default PaymentApiRequest;
+export default SingleImmediatePaymentRequest;
