--- src/models/providers/payments-api-response.ts
@@ -1,5 +1,0 @@
-import { Provider } from './provider';
-
-export interface ProvidersResponse {
-  results: Provider[];
-}
