--- src/models/providers/provider.ts
@@ -1,10 +1,0 @@
-/* eslint-disable camelcase */
-
-export interface Provider {
-  provider_id: string;
-  logo_url: string;
-  icon_url: string;
-  display_name: string;
-  country: string;
-  release_stage: string;
-}
